requirements.txt
requests
haystack-ai
sentence-transformers
torch
pypdf
numpy
pandas
scipy
scikit-learn
torch
torchvision
transformers
sentence-transformers
accelerate
tqdm
python-dotenv
pydantic
requests
bitsandbytes


accelerate==1.12.0
aiosqlite==0.22.1
alembic==1.18.3
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
Authlib==1.6.7
backoff==2.2.1
beautifulsoup4==4.14.3
bitsandbytes==0.49.1
blinker==1.9.0
boilerpy3==1.0.7
cachetools==6.2.6
canals==0.2.2
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cryptography==46.0.4
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
Events==0.5
exceptiongroup==1.3.1
farm-haystack==1.18.1
fastapi==0.128.6
filelock==3.20.3
filetype==1.2.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
google==3.0.0
google-adk==1.24.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.136.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.0
google-cloud-bigquery-storage==2.36.0
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.1
google-cloud-pubsub==2.35.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.62.0
google-cloud-speech==2.36.1
google-cloud-storage==3.9.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.62.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.1
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
haystack-ai==2.23.0
haystack-experimental==0.19.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
inflect==7.5.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-imports==0.3.1
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
mcp==1.26.0
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.16.0
networkx==3.4.2
num2words==0.5.14
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==2.16.0
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
platformdirs==4.5.1
portalocker==3.2.0
posthog==7.8.0
prompthub-py==4.0.0
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyarrow==20.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
pypdf==6.6.2
PyPDF2==3.0.1
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
qdrant-haystack==10.2.1
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-cache==0.9.8
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.2.0
sseclient-py==1.9.0
starlette==0.52.1
streamlit==1.54.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.5.2
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
torch==2.6.0
torchvision==0.21.0
tornado==6.5.4
tqdm==4.67.3
transformers==5.0.0
triton==3.2.0
typeguard==4.4.4
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
websockets==15.0.1
zipp==3.23.0
